+++ local.ts
@@ -1,0 +1,7 @@
+export let func = ():void=> {
+    console.log("hello flower")
+}
+
+export let namefunc = (val:string):void=> {
+    console.log("hello pakistan" + " " + val)
+}
