+++ local2.ts
@@ -1,0 +1,5 @@
+import { func, namefunc } from "./local";
+
+
+
+namefunc("Qadoos")
