+++ local3.ts
@@ -1,0 +1,3 @@
+import { namefunc } from "./local";
+
+namefunc("Gulraiz")
