+++ schoolsystem.ts
@@ -1,0 +1,3 @@
+import {TeacherOfDay } from "./teacherday";
+
+TeacherOfDay{}
